backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
alembic>=1.12.0
# psycopg2-binary>=2.9.0  # Only needed for PostgreSQL, optional for SQLite
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0
openai>=1.3.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
aiohttp>=3.9.0
pandas>=2.1.0
nltk>=3.8.0
dateparser>=1.1.0

